+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-26T10:43:25.088Z",
+    "2026-06-26T14:07:38.161Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {